Añadir codigo PlatformIo del sensor.
