+++ Codigo/ sensores/sensorUV_ML8511/SensorUV/.vscode/launch.json
@@ -1,0 +1,44 @@
+// AUTOMATICALLY GENERATED FILE. PLEASE DO NOT MODIFY IT MANUALLY
+//
+// PIO Unified Debugger
+//
+// Documentation: https://docs.platformio.org/page/plus/debugging.html
+// Configuration: https://docs.platformio.org/page/projectconf/section_env_debug.html
+
+{
+    "version": "0.2.0",
+    "configurations": [
+        {
+            "type": "platformio-debug",
+            "request": "launch",
+            "name": "PIO Debug",
+            "executable": "C:/Users/Alumno/Documents/PlatformIO/Projects/SensorUV/.pio/build/esp-wrover-kit/firmware.elf",
+            "projectEnvName": "esp-wrover-kit",
+            "toolchainBinDir": "C:/Users/Alumno/.platformio/packages/toolchain-xtensa-esp32/bin",
+            "internalConsoleOptions": "openOnSessionStart",
+            "preLaunchTask": {
+                "type": "PlatformIO",
+                "task": "Pre-Debug"
+            }
+        },
+        {
+            "type": "platformio-debug",
+            "request": "launch",
+            "name": "PIO Debug (skip Pre-Debug)",
+            "executable": "C:/Users/Alumno/Documents/PlatformIO/Projects/SensorUV/.pio/build/esp-wrover-kit/firmware.elf",
+            "projectEnvName": "esp-wrover-kit",
+            "toolchainBinDir": "C:/Users/Alumno/.platformio/packages/toolchain-xtensa-esp32/bin",
+            "internalConsoleOptions": "openOnSessionStart"
+        },
+        {
+            "type": "platformio-debug",
+            "request": "launch",
+            "name": "PIO Debug (without uploading)",
+            "executable": "C:/Users/Alumno/Documents/PlatformIO/Projects/SensorUV/.pio/build/esp-wrover-kit/firmware.elf",
+            "projectEnvName": "esp-wrover-kit",
+            "toolchainBinDir": "C:/Users/Alumno/.platformio/packages/toolchain-xtensa-esp32/bin",
+            "internalConsoleOptions": "openOnSessionStart",
+            "loadMode": "manual"
+        }
+    ]
+}
